requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
rembg[cpu]>=2.0.55
pillow>=10.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.1
boto3==1.29.7
